feat: rpc benchmarks
